PHOENIX-2503 Multiple Java NoClass/Method Errors with Spark and Phoenix
The calcite dependency in the regular client JAR is pulling in a version
of com.fasterxml.jackson which is incompatible with the Spark runtime.

This patch creates a new assembly artifact, client-spark, which attempts
to include only the client JARs necessary for the Spark integration to
work.

Also made sure the Phoenix driver is explicitly registered in the PhoenixRDD

Conflicts:
	phoenix-spark/src/main/scala/org/apache/phoenix/spark/PhoenixRDD.scala